Fix and refactor the rest (basically).
I'm sorry for this humongous commit, but there were too many interrelated
changes big and small to be able to break it apart in retrospective.
At the same time, currently everything seems to work reasonably well and I
haven't really touched the code for some time, so I commit this now as a
basis for further improvements. It should be in a consistent enough state to
be used, although still far from perfect.